// src/bot/operation/error/types.ts
export enum OperationErrorCode {
  UNKNOWN = 'UNKNOWN',
  ERROR_GETTING_ASSET_PRICE = 'ERROR_GETTING_ASSET_PRICE',
  ERROR_GETTING_AVAILABLE_BALANCE = 'ERROR_GETTING_AVAILABLE_BALANCE',
  MINIMUM_OPERATION_COST_LIMIT = 'MINIMUM_OPERATION_COST_LIMIT',
  ERROR_CREATING_BUY_ORDER = 'ERROR_CREATING_BUY_ORDER',
  ERROR_CREATING_SELL_ORDER = 'ERROR_CREATING_SELL_ORDER',
  ERROR_CREATING_STOP_LOSS_ORDER = 'ERROR_CREATING_STOP_LOSS_ORDER',
  ERROR_CREATING_EMERGENCY_SELL_ORDER = 'ERROR_CREATING_EMERGENCY_SELL_ORDER',
  BUY_ORDER_NOT_EXECUTED = 'BUY_ORDER_NOT_EXECUTED',
  EMERGENCY_SEL_ORDER_NOT_EXECUTED = 'EMERGENCY_SEL_ORDER_NOT_EXECUTED',
  SELL_ORDER_CANCELLED = 'SELL_ORDER_CANCELLED',
  STOP_LOSS_ORDER_CANCELLED = 'STOP_LOSS_ORDER_CANCELLED',
  STOP_LOSS_ORDER_FAILED = 'STOP_LOSS_ORDER_FAILED',
  STOP_LOSS_ORDER_EXPIRED = 'STOP_LOSS_ORDER_EXPIRED',
  ERROR_TRACKING_ORDER = 'ERROR_TRACKING_ORDER'
}

export type OperationErrorData = {
  [A: string]: unknown
  code: OperationErrorCode
}